리포트 로직변경 및 자잘한 정리

+++ src/main/java/qtedu/Impact_design/api/dto/response/report/ReportResponse.java
@@ -124,7 +124,7 @@ public class ReportResponse {
     @Builder
     public static class ScoreDetail {
         private Integer intentIndex;
-        private Integer stdntNo;
+        private Long stdntNo;
         private Integer score1;
         private Integer score2;
         private Integer score3;
