+++ src/main/java/qtedu/Impact_design/storage/jparepository/web/FLetterOfIntentJpaRepository.java
@@ -15,16 +15,16 @@ public interface FLetterOfIntentJpaRepository extends JpaRepository<FLetterOfInt
     List<FLetterOfIntent> findByCanvasIdIn(List<Long> canvasIds);
 
     // 사용자의 모든 투자 의향서 (삭제되지 않은 것)
-    List<FLetterOfIntent> findByStdntNoAndDelYn(Integer stdntNo, String delYn);
+    List<FLetterOfIntent> findByStdntNoAndDelYn(Long stdntNo, String delYn);
 
     // 특정 팀에 대한 투자 의향서
-    Optional<FLetterOfIntent> findByStdntNoAndInvestmentTargetAndDelYn(Integer stdntNo, String investmentTarget, String delYn);
+    Optional<FLetterOfIntent> findByStdntNoAndInvestmentTargetAndDelYn(Long stdntNo, String investmentTarget, String delYn);
 
     // 특정 팀이 받은 모든 투자
     List<FLetterOfIntent> findByInvestmentTargetAndDelYn(String investmentTarget, String delYn);
 
     // 제출 완료 여부 확인
-    boolean existsByStdntNoAndSubmittedAndDelYn(Integer stdntNo, Boolean submitted, String delYn);
+    boolean existsByStdntNoAndSubmittedAndDelYn(Long stdntNo, Boolean submitted, String delYn);
 
     // 내 팀을 투자 대상으로 한 모든 투자 (F3용)
     List<FLetterOfIntent> findByInvestmentTargetAndDelYnAndSubmitted(String investmentTarget, String delYn, Boolean submitted);
